+++ Chromagnon/__init__.py
@@ -6,12 +6,12 @@ if sys.version_info.major == 3:
 __author__ = "Atsushi Matsuda"
 
 try:
-    from . import version, cutoutAlign, alignfuncs, chromformat, aligner, chromeditor, threads, flatfielder, chromagnon
+    from . import version, cutoutAlign, alignfuncs, chromformat, aligner, chromeditor, threads, flatfielder, extrapanel, chromagnon
 except (ValueError, ImportError):
     try:
-        from Chromagnon import version, cutoutAlign, alignfuncs, chromformat, aligner, chromeditor, threads, flatfielder, chromagnon
+        from Chromagnon import version, cutoutAlign, alignfuncs, chromformat, aligner, chromeditor, threads, flatfielder, extrapanel, chromagnon
     except ImportError:
-        import version, cutoutAlign, alignfuncs, chromformat, aligner, chromeditor, threads, flatfielder, chromagnon
+        import version, cutoutAlign, alignfuncs, chromformat, aligner, chromeditor, threads, flatfielder, extrapanel, chromagnon
 
 reload(version)
 reload(cutoutAlign)
@@ -21,6 +21,7 @@ reload(aligner)
 reload(chromeditor)
 reload(threads)
 reload(flatfielder)
+reload(extrapanel)
 reload(chromagnon)
 try:
     from . import testfuncs